Featherhat (#101)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added in a quick cosmetic item; a feathered cap! Sprite is done by Lore
on the discord. I've put it in the loadout section for hats, as well as
the mail package loot table.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Added the Cap

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![worn](https://github.com/user-attachments/assets/b167d8e6-9c29-4c0c-92c0-61a73c1f539e)

![worn2](https://github.com/user-attachments/assets/e5e0b4c4-6775-4dbe-8a63-9b78c788faa1)

![the](https://github.com/user-attachments/assets/0958feec-7a38-432a-9350-527223f8abd5)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Feathered Cap