Alice Plushie (#105)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds a plushie of Captain Alice Maranna from the NT "Cuddle Command"
propaganda plushie line.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: "Cuddle Command" Plushie # 1 Alice Maranna

---------

Co-authored-by: juniwoofs <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>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>