Fix: Server Now Reads Request Grid Rather than Location Grid (#132)
When originally coded for frontier both the target grid and the location
grid were the same. i missed that the server was not reading the grid
sent in the request but rather the grid the console was sitting on. on a
normal station this manifests its self by making it look like the cruise
control is not working, however what is going on is that you are
changing the dampening values of the station, not the shuttle. oops

I also hate the names they changed the dampeners to, but i am to lazy to
go change those :) enjoy

:cl:
- fix: Fixed inertia dampeners

Co-authored-by: neuPanda <spainman0@yahoo.com>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>