Added the Argenti, Gestio, Novalite, SVT, and Harpoon to the Salvage Vend (#133)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds various weapons into the salvage vending machine, such as the
Argenti revolver, the Gestio, Novalite, and SVT rifles, and a new spear
type Harpoon weapon for salvage.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added a revolver, three rifles, and a harpoon into the salvage
points vend, the guns do not come loaded for workplace gun safety. Have
a fun time hunting your white whale!

---------

Co-authored-by: Rosycup <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>