Cargo Shuttle Fixed (#134)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

Howdy y'all, I noticed the cargo shuttle blast doors only open, not
close. I'm going to fix that. I realigned them correctly ~~with the
airtight plastic flap and~~ with Tiny fans fixed the problem with the
air system. There was only air coming in, but no waste air out. I added
two chairs and some tools.



https://github.com/user-attachments/assets/4615ebb1-12aa-4c96-a2c4-d52d898bbee3


---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)
![Screenshot 2025-01-06
012814](https://github.com/user-attachments/assets/c0af2571-9469-43c6-98f5-14b2d3d9e693)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Mike32oz
- add: Added an air alarm and sensor additional tools.
- tweak: Tweaked cargo shuttle blast doors
- fix: Fixed cargo shuttle Air/Waste distro
- remove: Removed tippy the clown (my nemesis) airtight plastic flap