Pebble Violet V2 (#162)
# Description
Howdy, y'all. This is a quick update for Pebble Violet Edition. I just
added the station anchor/AI and uploaded computer with the law boards.

Some minor fixes of wiring and spawnpoint/cryo


<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)
![Screenshot 2025-01-17
094103](https://github.com/user-attachments/assets/a696fc7b-6236-4ca8-bcef-ea5d81c84b09)
![Screenshot 2025-01-17
094038](https://github.com/user-attachments/assets/bcb36b6d-6cce-4066-afd7-75c061603d4d)
![Screenshot 2025-01-17
094013](https://github.com/user-attachments/assets/9a6781c5-8215-47f0-a069-337187b9a7b2)
![Screenshot 2025-01-17
094008](https://github.com/user-attachments/assets/0da4877e-ec63-485f-8d05-3b4d3643e48d)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Mike32oz
- add: Pebble, Station anchor.
- add: Pebble, Station AI with AI upload Law boards. 
- tweak: Cable management, Cryo late join spawner