Pebble Violet V4 (#173)
# Description


Howdy, y'all I just added the spawn points for the new Magistrate. I
also fixed a bug that was causing the player to get stuck on the wall in
the Medical Cryo area. Finally, I made minor fixes to the Atmos
Department airlocks, which were not locking correctly, and added an APC
to the TEG.


<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)
![Screenshot 2025-01-19
041049](https://github.com/user-attachments/assets/b7e72a27-6ecc-48d1-9d2f-5a34484497c9)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Mike32oz
- add: Pebble, Added Magistrate spawn job
- tweak: Pebble, Atmos airlocks, Medical Cryo
- remove:  Tippy the Clown (my nemesis)

---------

Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>