Big Baikal Fixes! (#193)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Violet <3 and I fixed the station (some of it)

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Violet and I sat there for ~3-4 hours and worked down the list. This
included a few QOL changes, some modifications to the map, and some
further progress.

---


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: juniwoofs and Mike32oz 
- add: Added the Prideomat (being gay is now legal!)
- add: Added a reagent grinder to chem! (have fun blending things!)
- add: Added a drain to disposals (now you wont have the infinite slop
puddle)
- add: Added a seed vendor to perma (finally, food for the poor)
- add: Added a lot of stuff to maints (a few rooms, loots, and
water/fuel tanks)
- add: Added the gas miners the new mapping tools took from us (no
longer 1984)
- add: Added a few things of loot to the outside of the station
- add: Added curtains to perma (you have a lil privacy now)
- tweak: Tweaked the cargo shuttle dock to make it a little easier to
dock the shuttle
- tweak: Tweaked epi's front door (better placed now, trust me)
- tweak: Tweaked engi's air alarms! (no more dying when the air is spicy
thanks mike!)
- tweak: Made engi generally better (thanks also mike!)
- fix: Fixed Captain locker being empty (now you can have your fancy
tools and dont have to share the two sticks and a rock with the whole
platoon)
- fix: Fixed the boxing room floors giving you xray vision. (the
concrete may have been bought from less reputable sources...)
- remove: Removed cameras in the perma dorm rooms (sex is legal again
for the incarcerated)