Blueshield Rescue Axe (#210)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds the Blueshield Officer's Rescue Axe, both a personal one-handed
self defense weapon and a prying tool designed to pry open doors in the
event of a low power emergency within the station in order to reach the
Blueshield's assigned VIP.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added the Blueshield Officer's Rescue Axe, both a personal weapon
and a tool used for emergencies issued to Blueshields as part of their
equipment.