Nicotine Addiction Tweak (#213)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

A small tweak to the nicotine addiction to be slightly less punishing
(especially with how easy it is to get), serves as a bandaid fix until
someone on upstream improves on the addiction system. Not smoking for
some time causes the initial mood debuff to degrade into a weaker mood
debuff, although still permanent for balance.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Tweaked nicotine addictions to degrade into a weaker mood
debuff with time spent not smoking. May your recovery be full of
progress.