Melee Tweaks and Changes (#289)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

**Melee Tweaks**
- Greatsword item size increased, added movement speed debuff when
equipped on back, higher stamina drain on wide swing.
- Circular saws don’t have the silly 360 swing animation.
- Stun Batons take 5 stamina to wide swing (increased from 1 stamina).
(A metal baton like that should NOT be that easy to swing)
- Mining Equipment Tweaks
 	- Pickaxes take 7.5 stamina to wide swing.
	- Crusher Axes now deal 13 Blunt and 10 Slash on wielded attacks.
	- Crusher Glaives now deal 13 Blunt and 15 Slash on wielded attacks.
	- Crusher Daggers now deal 12 Slash.
- The Crusher mark bonus is increased from 20 Blunt and 5 Slash to 25
Blunt and 10 Slash.
	- Drills now take 0 Stamina to wide swing.
- Shivs are now slightly slower
- Katanas deal 17 Slash upped from 12 and now take 7.5 Stamina to wide
swing, half of its original stamina cost.
- Cutlasses deal an additional 3 Piercing on top of its original 12
Slash
- Katanas and Claymores are now slightly faster at wide swings.
- Captain Sabre damage upped from 15 Slash to 18 Slash, reflect chance
upped from 10% to 20%. You better not be charging in with this thing.
- Energy Bolas are now 2x1 items, fit into security belts, and deals
slightly more slow.
- Fireaxes, Kanabous, Breaching Hammers, Baseball Bats, Crusher Axes and
Glaives, and the Guitars are now able to fit on both back and suit
storage slots.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: A whole bunch of melee tweaks, view the PR for complete
details.