Glacier II Update (#287)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

- Adds holopads
- Adds private dorms
- Adds arrivals bay
- Extends salv dock
- Actually gives Sec guns
- An extra pride o' mat. Fuck  you, get gayer.
- also adds a silo i guess

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Fox
- tweak: Glacier II has been updated with holopads and more, have fun in
the dorms you filthy animals