Who Up Shoukouing They Shit (Proper Spelling This Time) (#291)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description
Ports the Delta-V amendments to Shoukou, maps in magistrate and NT rep
and Blueshield. Blueshield lives in the bathroom of the bridge bar,
NT-Rep lives in the place where Clerk was, CJ became Magistrate.
Secret rooms that Juniper built are intact. 

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

There is no media this time, refer to previous PR for the one image I
have of this.

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Shoukou has now been updated to a variant of its Delta-V
version. It has retained its cozy rooms from the previous version.
Science was moved, we now have a library, go have fun.
- add: Added Pluey.