Wisps and Fixes (#309)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Glimmer Wisps returning, Mites now have a lower minimum glimmer
requirement to spawn (From 700 to 620),Localization fixes for the vent
events, Luminous creatures spawned by the flora anomaly are now psionic
to fall in line with the rest of the anomaly based creatures, Reagent
Slime Bottom Surgery.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Glimmer Wisps have returned.
- tweak: Luminous Entities now count as a psionic creature, to match
with the rest of the anomalous entities.
- tweak: Gave certain reagent slimes bottom surgery, good for them.