Arena Update (#319)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

- Expands burn chamber and removes atmos piping (most of it)
- Adds fans to the arrivals arm
- Makes all dorm rooms AA and gives them locks
- Adds BSO and NTRep lockers
- Minor lighting fixes

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Fox
- tweak: Arena has been adjusted a little, theres like 8 dorms rooms
now.