Quick Marking Adjustments (#322)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Fixed corgi tail typo, Diona overlay and one more point for chest
marking for IPCs to prevent disappearance when picking a diona-based
chest marking.

---