Humanhybridchanges (#327)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->
I made more felinid markings available to human hybrids and slimepeople,
as well as creating furry limbs for those who aspire for a more
'kemonomimi-ish' appearance.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- I dunno man I just put funny words into the boxes

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media (CW: naked cat [actually human
hybrid])</h1></summary>
<p>

![SPOILER_image
1](https://github.com/user-attachments/assets/a7b75bae-1d2e-4ada-9a31-0cd9190e1c02)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: furry limbs for human hybrids
- tweak: human hybrids and slimepeople can now have different types of
cat ears and tails