Tortuga Update (#329)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

- New Dorms and a new hallway between arrivals and logi
- Total atmos death
- Minor lighting tweaks
- Minor power cable tweaks
- Fixed* salvage dock bug

*maybe

---


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Fox
- tweak: Tortuga has been updated! All Den Maps in rotation are up to
standard!