Sundry Hydroponics and Food Stuff. (#337)
Description
Adds an addition plant, seed and recipes and (and associated flavours) -
including a recipe for sausage. None of the additions have been added to
the the entity spawner, however, so will not be randomly spawned on
maps.

It all seemed to run well on my end, but could probably use a look-over.

Apologies if this isn't up to standard, this is my first ever attempt at
contributing!

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: An additional plant, seed, recipes and flavors, including a
recipe for sausage.

---------

Co-authored-by: Hederadaii <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>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>
Co-authored-by: Rosycup <178287475+Rosycup@users.noreply.github.com>