Humans Have Human Hybrid Features (#336)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

As was ordered, so shall it be. This PR attempts to consolidate Humans
and Human Hybrids into effectively the same thing. As per this PR, any
markings that were previously restricted to Human Hybrids are now
available to humans. Additionally, there is a new skinColoration that
allows a species to toggle between using RGB/HSV sliders and the human
skin tone slider. At this point in time, there should be nothing that
human hybrids have that humans do not. With this, there should be no
issue should human hybrids be deleted, though I will make that PR later
as to give warning. Another minor change is that the 'cat ears' and 'cat
tail' markings have been removed, as they are the exact same as the
basic felinid ears and tail except they default to hair color instead of
skin color.

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![newbutton](https://github.com/user-attachments/assets/e3dfee7e-0064-4ecb-9e4e-41ec09478d20)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: new skinColoration that allows a species to use both RGB/HSV and
human skin tone sliders, available as a toggle.
- tweak: Humans should now be able to use anything that Human Hybrids
can
- remove: the cat ears and cat tail have been removed