Return Of The Hive (#334)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

- Strips atmos pipes from main atmos
- Does NOT touch SM Chamber atmos
- Spawnpoint adjustments
- Wiring adjustments
- Lighting adjustments
- If you break the SM its not my fault
- I cannot stress this enough, be careful with the SM
- Please
- Changes evac shuttle to the Vertex to hopefully avoid an RTA with the
arrivals shuttle

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Fox
- add: The Hive Returns. Please, do NOT break the Supermatter. I will
cry.