Blueshield Shield and Mace (#339)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

What if we gave BSO more weaponry, that'd be funny and cool I thinks.
Cherry-picked from Goobstation
(https://github.com/Goob-Station/Goob-Station/pull/1603)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Add the Greatshield & DT-3 Atrocity
- [x] Add them to loadouts
- [x] No longer from master

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/91815915-d269-42b0-961a-e8a3b635b1dd)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: SolsticeOfTheWinter (and KyuPolaris for tweaking to TheDen)
- add: More warcrimes for BSO to commit. Also known as a Greatshield and
Mace

---------

Co-authored-by: Solstice <solsticeofthewinter@gmail.com>
Co-authored-by: KyuPolaris <tabi@kyuudev.me>