Shoukou Hotfix (#349)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Mono finally fixes the Shoukou docks and makes the arrivals identical to
evac, makes a decent pathfinder parking spot, and nukes the docks at the
shipyard area, replacing them with external doors with this one weird
trick. Runs locally.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/a6d32759-dadd-402b-9966-93bf7332d8e7)
It is not actually named Evil Fucking Fox.

![image](https://github.com/user-attachments/assets/ee9e75d7-2925-4ae8-a1e9-27cd7dcf673f)
Reference image for how the pathfinder parks which was used for the
shape of the one that docks. I hope the salami lid fits.

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: Tweaked all of Shoukou's docks so that arrivals only goes to evac
and arrivals in theory. Also, makes sure the salami lid of the
Pathfinder fits, in theory.
- remove: Removed Pluey.