Corpsman Summer Gear (#356)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Now the Corpsmen won't melt in the summer heat.


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Corpsmen Summer Uniform, now you won't die to space summer in
your stuffy uniform!

---------

Co-authored-by: Rosycup <178287475+Rosycup@users.noreply.github.com>