Honey Stomach for Chitinids (#358)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds a Honey Stomach trait to Chitinids, this trait works similar to the
lewd organ traits and produces honey for the cost of hunger.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Honey Stomach C# Code
- [x] Honey Stomach Trait
- [x] HoneyProducer localisation file

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

![image](https://github.com/user-attachments/assets/fd6f79ae-92d4-4652-aef0-d7bb3d28ba95)

![image](https://github.com/user-attachments/assets/277becab-d948-467e-9d58-eb70997d2bf8)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added the honey stomach organ for chitinids

---------

Co-authored-by: KyuPolaris <tabi@kyuudev.me>