Quickfixes (#362)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Water Vapor Canister added to the cargo catalog following Feroxi
species, added meat tags to one of the new sushi variants, New Ammonia
and Tritium canisters sprites.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Water Vapor canisters to the cargo catalog.
- add: New Ammonia and Tritium canister sprites.