Adjusts Jobs Loadouts to Make Them More Uniform (#375)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

A simple reorganization and relaballing of certain items, only ONE of
which has been removed (commented out) (as the item its associated with
does not exist), collars are now accessable by the whole department
rather than just some arbitrary jobs inside of it and some items have
moved into the correct categories

Also adds the jumpsuits and jumpskirts of the default uniforms that a
job has, those being for, corpsman, mime, security officer, warden, head
of security, botonist, bartender and chef (i like using default
jumpskirts :3c)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Changes made
- [x] Tested

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Loadouts have been reorganized please make sure you update
yours!