Quicktweaks (#383)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Gave Blueshield Officer a pinpointer in their locker, Moved Blueshield
Axe to loadouts, Dragon meat can now be crammed into custom burgers and
tacos.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Gave Blueshield Lockers a pinpointer.
- tweak: Blueshield Axe moved to loadouts.
- tweak: Dragon Meat available for custom burgers and tacos.