Canid (Reject Catgirl, Embrace Dogboy) (#409)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

A species much like the Felinids are to the Tajara, but for the
Vulpkanin instead, by request from the fox here they are.

All added locales are written by Fox, as are the nameslists (I tacked
Laika to the end of the female one)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [X] I did all of this
- [x] In like 6 hours
- [x] already tested it all, it works

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/3cb983b7-b835-4d1e-8a52-ead29aa427dd)

![image](https://github.com/user-attachments/assets/61892cb8-6b70-4023-9eb2-4e40746e6b59)

![image](https://github.com/user-attachments/assets/a10776ff-6f41-410b-a404-f6b53bd5d833)

![image](https://github.com/user-attachments/assets/f8bd3967-782b-4ac3-944d-6673a9d17f61)

![image](https://github.com/user-attachments/assets/dd4b0c90-48eb-445f-ab20-95f8e948d325)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: BlitzTheSquishy, Fox
- add: CANID! Heavily gene tailored humans to have dog/fox-like
features.