Shoukou Minor Changes (and Lavaland I Guess) (#428)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Does a couple things:
- Adds vents to the BSO's office because one BSO main requested it,
since they need to be able to breathe you know? I can't believe CC
didn't even have the courtesy to provide air to the bathroom that was
there before
- Adds sec space suits. There's three of each delta-v variant (advanced,
normal, riot) in the armory now. When you need them they are there. HOS
and Warden suits not included because I think they're already packed in
those lockers
- Adds a note of courtesy to the atmos fucking room since the entrance
to that isn't secure and someone could easily interrupt a scene,
hypothetically, if they entered without knowing someone was there.

When I pull master to this branch because it doesn't seem to be
functioning on my device and I don't have the latest version:
- Extends the salvage dock to accomodate lavaland shuttle
- Maybe adds an extra room to Engi because there's starting to be a lot
of dead space on that dock and I wanna take up a little bit of that. Not
sure.
- Tweaks the pathfinder dock because according to juni it was "too
tight"
 ---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

If it's necessary I will edit this, but the text basically covers it and
most of this is minor enough to not warrant a picture

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Adds Lavaland's shuttle to Shoukou.
- tweak: Edited Shoukou Security to give it space suits. The days of
evil salvage contraband criminal times are over, it's time to go to
jail.
- tweak: Shoukou has a new (and not ass) emergency shuttle.
- fix: Fixed the BSO not being able to have sex and the pathfinder, much
like the salami lid, not being able to fit on the station.