Asterisk Face-Lift (Greggified) (#431)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This is a PR about Asterisk, Asterisk was made by my favorite Mapper,
Colin_Tel. 😄 _(I was held at gun point to write this)_
Real talk, I wanted to give Asterisk the adjustments it needed for use
on the Den in specific. Essentially, added two more dorms, gave all
dorms Atmos and ensured the Dorms have linked buttons for door bolting
and enough space between bed to outer areas so nobody can hear Subtles.
Also added Anti-Anomaly Generation zones around Dorms/Cryo, to ensure
minimal scene disruption.
Beyond that, we now have on Asterisk the BSO's office, NT Rep's office,
an Admin Assist desk in Command, Document Printer in the Conference
Room, an airlock leading into the Burn Chamber in the TEG room, a
Lavaland Shuttle dock with interior access into Salvage so you don't
have to wear suits rated for space-wear, moved Janitorial, added a
Material Reclaimer to Disposals, gave Security cells to use that aren't
in Perma at the cost of the Bomb Defusal training area, added three
Security Hardsuit storages inside Security next to the Locker Room, a
Soda Dispenser in the Maints bar w/ LV wiring so it's powered from round
start and **LAST BUT CERTAINLY NOT LEAST** I attempted to remove the
wasteland under the bar floor in Asterisk. I say attempted to, because
when attempting to pry up the floor tiles in the Mapping Environment, it
did not replicate the issue as it would on Live with Wasteland tiles
being beneath the floor. Unsure if this actually fixed anything, but
hopefully it should. Replaced the wooden tiles with the standard wooden
tiles.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Possibly give the Maints bar some TLC to make it easier to setup
and operate.
- [x] Gave Asterisk a once over. Will very likely come back over it for
another pass as time goes on.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Expanded dorms on Asterisk
- tweak: Made various QOL improvements to Asterisk
- remove: Removed whimsy >:(