Vocal Dionae, Take Two (#437)
# Description

My dumbass forgot to set the right branch the first time. Brings over PR
#32511 by Pinkbat5 from WizDen, which (if I did this right) will allow
for the whole three (3) Diona players on The Den to be able to use the
chirp emote while in their adult Diona form, and allow for Nymphs to
have the same screaming and laughing SFX as the aforementioned adult
form.

---

# Changelog

:cl:
- add: Dionae can be noisy now! Added the chirp emote and associated
sound for Diona, along with screaming and laughing sounds for Nymphs!

Co-authored-by: pocl v <24708225+Pinkbat5@users.noreply.github.com>