Glacier QOL (#450)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Changes made with @wallflowerghost
- Changing rooms to both saunas
- LV Wires added to salv dock
- Lighting changes
- Moves Ian OUT of the Turbo Death Chamber 9000 Ultra Max Pro Deluxe
Edition with new Funky Mode
- Minor exterior changes
- "The Box"

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Fox & Wallflower
- tweak: Adds a bunch of QOL stuff to glacier. The sauna is USABLE now!