Leopard Tail for Tajarans (#444)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

There is no point making it long. It adds another awesome tail marking
for Tajrans, which is leopard tail. Big fluffy and with tail wag
animation. Its taken from Skyrat-TG. It can be found on all its
downstreams, but was not able to find commit which added it. Might be
old as bayrat.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Animations when dream maker starts to work

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/83aac3c6-fc8c-405d-914b-0c970a2e65fe)

![image](https://github.com/user-attachments/assets/2229cb03-5b2d-4c2a-ad98-b72ef17f6956)

![image](https://github.com/user-attachments/assets/5e519c10-3801-4855-a547-fa2c41169c81)

![image](https://github.com/user-attachments/assets/4287a450-77e4-4d01-ac34-bdce33f6f699)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added big fluffy leopard tail, exclusive for Tajarans.