Fixing Asterisk Part Three: Soju Vision Quest (#457)
Today, greggy boy sat on his goofy stool while overlooking the swamp and
spoke in tongues until things were maybe fixed or something.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

hey so like, I really just fucked up last time and put the updated map
in the wrong spot hehehe. I now have fixed that! Yay!
This time around, I just added lattice around the maints bar and
curtains so it's extra private and cool _puts on sunglasses_
but uh yeah, aside from that fixed hopefully the bar lights not being
powered due to the mounts being bugged out and I fixed the Station
Beacon for the bar being unanchored and above the tiles.
P.S. to future and sober gregg, fuck you dipshit. put your map files in
the right spot or we're getting sent back to the shadow realm.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Uploaded the Map file correctly this time!

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added malice and pestilence
- tweak: Tweaked whimsy spawn rates.
- fix: Fixed where I uploaded my map files