Fixes and Tweaks (#460)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Rat King and Servant AI fixes, Theoretically should fix the sol
deserters, reverts the vehicle code on office chairs for the mappers.
Fixes whitelisting for empty pistols. Added medical masks to the
Nanomed. Added Ice cream bowls to lathes.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: Fixed rat servant AI.
- fix: Fixed empty guns having the wrong magazine whitelists.
- tweak: Office Chairs reverted to normal chairs for mapping purposes.
- tweak: Gave sol deserters their own chest rig variant to fix issues.
- add: Added the medical masks to the Nanomed.
- add: Added the ice cream bowls to the autolathe.