Psionic Insulation Tweaks (#461)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Tweaks the waveform misalignment trait text on the loadout screen, adds
a new examine text for it as well.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Grip someone firmly and get them to test the PR for me to see if
Psionically Insulated people still gain powers and if the mass sleep
immunity works. (Tabi did this and deserves a kiss)
- [x] Do the PR

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Tweaked the trait flavor text for Waveform Misalignment and the
examine text for it.