Ethanol Content Fix & Aphro Tweaks (#447)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This will be changing the ethanol values of alcohols and cocktails
within the game, to make it more in line with real values. I'll change
the Libidozenithizine recipe while I'm here, as I think I made it far
too expensive. In line, all of the Aphros will have their metabolism
rates readjusted, and ODs.
---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Base Drinks
- [x] Nyano Drinks (some still exist)
- [x] Delta-V Drinks
- [x] Floof Drinks
- [x] LBZ recipe
- [x] Aphrodisiac Tweaks
---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: SleepyScarecrow
- tweak: Tweaked Ethanol of Cocktails
- tweak: Tweaked Libidozenithizine recipe to be less expensive
- tweak: Tweaked Aphrodisiacs OD, Metabolism, and Message Rates

---------

Co-authored-by: SleepyScarecrow <sharedjunkemail69@gmail.com>