Blueshield and Admin Assistant Additions (#464)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

I will be surprised if I don't go several YAML Linter test fails from
this. Adds a secondary weapon loadout category, Fountain Pen, and Rubber
Stamp for Blueshield, replaces the headsets for Admin Assistant and
Blueshield to the over ear variants. Updates the Admin Assistant ID
overlay. Ports the Administrative Assistant locker from Delta-V, up to
the mappers if it gets mapped.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Port Admin Assistant Locker from DV
- [x] Small stuff

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added rubber stamp and fountain pen for Blueshield Officer.
- add: Ported the Administrative Assistant Locker for mapping.
- tweak: Tweaked BSO and Admin Assistant headsets to the over-ear
variants.
- tweak: Gave a new sprite for the Admin Assistant ID symbol.

---------

Co-authored-by: Radezolid <snappednexus@gmail.com>