Pebble Overhaul Part 1: "Jenn's Great Idea" (#465)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

- Adds Firing Range, Classroom, and smallerizes the court
- Moves detective office
- Minor QOL
- Gives Gray his damn books
- Prototype Changes

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Major work has been done to Pebble's south side. Plan a lesson in
the new classroom!
- remove: Removed Herobrine