Tortuga Overhaul (#468)
Attempted to give Tortuga the tender love and care it's needed for a
long time. Public facing APCs, an Epistemics re-work, better dorms
across the station for privacy, AI Sat, Holopads, tweaked the Armory and
general QoL changes.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->
1: Adjusted the Armory's safes and interior, hopefully making it less of
a pain in the dick for Warden's to work with.

2: Perma now has it's own APC.

3: A variety of APCs have been added or replaced. Bar, Boxing, Arrivals,
Arrival Dorms, Perma, Psych's office, etc. Lots of fucky wiring work was
pre-existing, so I know some things might not be wired to some folk's
specifications but please understand I'm not stripping ALL the wires,
Substations and APCs right now to fix this.

4: Ore Silo has been moved from the Engineering Material room to the
Vault.

5: The Engineering Material room has been turned into a Tech Vault.

6: Holopads have been added everywhere I could find a place for them.

7:The NT Rep's Office was shoved in the Chicken room. Sorry pal.

8:The BSO's Office was shoved in the Security Checkpoint right at the
Bridge entrance. At least I gave them a couch.

9: All the dorms have been tweaked to be more private, have more room in
them and/or give people creative room to tweak
things to their needs/scenes. The Maints rooms that previously needed a
disposal chute are now just airlocks and are very nice rooms to spend
private time in.
_But Gregg, those rooms are nicer than anywhere else on the station, it
doesn't make sense!_
Yeah. I honestly am beyond giving a shit, you can be a glowing red space
terrorist and have a pie thrown in your face. None of this makes sense.
People should be allowed private, accessible and comfortable things once
in a while, hence easy round start apartment-esque rooms. The beach room
has also been expanded, and extra things to make the room more
comfortable added as well. Extended latticing is now around that side of
the station.
 
10: AI Sat has been built, extending out past the Conference room at
Bridge. The AI has been removed from below the Armory.

11: Massive Epistemics changes.
The Epistemics Artifact lab has been tweaked. There is now a single
Artifact chamber, instead of two. The shared surgery room has been
tweaked, it is purely Medical's room now. Robotics has been given
additional space where a previous artifact room was. An empty room that
can easily be turned into an Anomaly Lab is available where the second
Artifact chamber used to be. The prober was moved slightly left, instead
of being directly in the Mysta's office. A circuit imprinter now is
available at round-start.

12: Admin Assist locker was added.

Seriously, if you find any issues with this, post them in the Tortuga
Mapping Forum and ping me. Comments, questions or concerns. This is
merely a first draft, this is not a locked in final product. Things can
and will be adjusted as time goes on.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Possibly take a look at the fuck-ass wiring.
- [x] AI Sat, BSO, NT Rep, Holopad, APCs added.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Tortuga Overhaul, view PR for details.