Quality of Life Additions and Changes (#494)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Created a racially specific language for Rodentia called Rodentian,
currently a direct copy of mouse in terms of noises made. Also changed
font for Psychomantic to the OpenDyslexic typeface as having an easily
read typeface is good for the universal language
---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] the thing 👍

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![Medal_8S62ZMKm1x](https://github.com/user-attachments/assets/056e0150-8ef1-41e3-8393-379cee355865)

![Medal_WUWpd3qDGh](https://github.com/user-attachments/assets/2cecf654-e562-4e75-9e37-fe7a8cdfdc67)
(the universal language will have [Psychomantic] before it, I forgot to
get screenshot)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Rodentian, a language based off of mouse specific to Rodentians
(they can no longer shittalk Hamlet)
- tweak: Psychomantic now uses the OpenDyslexic font to increase
readability

Co-authored-by: KyuPolaris <tabi@kyuudev.me>