Entity Menu Cleanup (#489)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

EE made things weird with how entities are shown inside of the entity
spawner menu, when they shouldnt be (namely events) this plans to fix
that.
Also adding tags to items with the same/similar names for easier
differentiation

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: BlitzTheSquishy
ADMIN:
- fix: removes items from the admin spawn menu that should not be
spawnable (commands uneffected)