Revert "Cyber Eye Changes (#1876)" (#498)
This reverts commit 516c53e8ea97680b8029b79dae9fab0a229d0709.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Reverts the flash resistant eyes back to existing (though slightly more
expensive to take in line with outer eye upgrades)

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Readded flash resistant cybereyes