Armor and Weapon Tweaks (#496)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Buffed Swat Armor to be proportional to its status as an expensive armor
set only obtainable through logistics.
Resistance values changed as follows:
- Blunt: 40% -> 55%
- Slash: 40% -> 55%
- Pierce: 40% -> 55%
- Blunt: 15% -> 50%
- Explosion: 50% -> 70%

Swat shoes buffed as well to have 60% slowdown decrease and 35% stun
time decrease, lying just above combat boots in terms of stats.
Gave the HoS Samurai Armor a 5% movement decrease and 40% explosion
resistance.
Made the Terminus and the RPG equippable on back and suit storage slots.
RPG requires wielding now to fire.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Buffed Swat gear to be a bit more proportional to its expensive
status and to compete with other riot gear better, check the PR for the
values.
- tweak: Tweaked HoS' Samurai Armor to have a smaller movement decrease
and explosion resistance.
- tweak: Terminus and the RPG equippable on back and suit storage slots,
RPG requires wielding to fire. You must use both hands to end a life.