Cowbell Collars (#504)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added cowbell collars for that can be used for 500 hours of mind pumping
action and all your livestock needs.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

![icon](https://github.com/user-attachments/assets/a0f9c053-63e0-43bb-a7c3-c9ecb9329515)



![equipped-NECK](https://github.com/user-attachments/assets/f794d034-f967-450a-a9e9-eda76668d7b0)

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added cowbell collar

---------

Co-authored-by: Rosycup <178287475+Rosycup@users.noreply.github.com>