Rodentia, Chitinid, FastMetabolism Hunger Fix (#508)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Ports https://github.com/Fansana/floofstation1/pull/897 and also fixes
your FastMetabolism trait. Also locks Vulps out of Fast Metabolism since
they already have accelerated hunger.

For reference, DeltaV (originating fork of the races) has a much higher
baseDecayRate for hunger. Rodentia, Chitinid, and Fast Metabolism all
had a ~180% increase instead of 33%. Go figure.

Additionally makes a tweak on Diona thirst so it's not insanely slow
based on discussion had in the PR.

This is a very silly thing I've done and I haven't even played on the
server yet.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Cherry-pick https://github.com/Fansana/floofstation1/pull/897
- [x] Adjust Fast Metabolism
- [x] Adjust Diona thirst decay

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

Eh

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: Chitinid, Rodentia, and those with Fast Metabolism now get hungry
at the intended rate. You poor, hungry souls.
- fix: Diona now get thirsty at the intended rate. They are not cacti.

---------

Co-authored-by: Rosycup <178287475+Rosycup@users.noreply.github.com>