Potential Pill Test Error Fix (#511)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Pills are designed to do some kind of networking and serialization for
their visuals, for reasons I do not comprehend (maybe the chemmaster?),
via the PillComponent and PillSystem. As part of this, the entity system
is setting the pill state during "AfterAutoHandleStateEvent" (another
thing I currently do not understand) to a semi-hard-coded value of
"Pill<num>" where <num> is the PillType defined in the component
(defaulting to 0) + 1. Something in the tests is triggering this
"AfterAutoHandleStateEvent" event, attempting to set the state of phil's
popup to pill1 (a state which did not exist), and then throwing errors.

Anyway, this proposed fix is likely both the cleanest and dumbest way to
fix the consistent errors if I am correct. Feel free to
reject/close/propose changes if you don't like the solution, but I hope
this helps.


---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Spend a bunch of time looking into a minor issue on a server I
still haven't played on because I apparently can't help myself

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

No CL :(