Cherrypick: Message Highlights (DeltaV#2990) (#506)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Ports PR [#2990](https://github.com/DeltaV-Station/Delta-v/pull/2990)
from DeltaV which adds highlighting to chat using certain keywords. It
helps neurodivergent people and otherwise easily distracted people be
able to pay attention when their names are called a lot easier.

It also adds a tab to the menu "Extra" that contains all the options for
the feature

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

It's about as far as I'm going to take it, if someone wants to port the
color picker from DeltaV, be my guest

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/808fb983-8587-449c-b803-7d4989ad0372)

![image](https://github.com/user-attachments/assets/312e8f8e-d3f9-48b3-8354-691376ada6e5)

![image](https://github.com/user-attachments/assets/da51d1ac-60e2-4a72-93c5-4849f83d2030)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Ported chat highlighting from DeltaV

---------

Co-authored-by: Tobias Berger <toby@tobot.dev>
Co-authored-by: V <97265903+formlessnameless@users.noreply.github.com>
Co-authored-by: Rosycup <178287475+Rosycup@users.noreply.github.com>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>